img_caption_with_vision/requirements.txt
google-cloud-aiplatform==1.28.1
google-api-core==2.11.1
google-auth==2.22.0
google-cloud-core==2.3.3
google-cloud-storage==2.10.0
google-cloud-vision==3.4.4
httplib2==0.18.1
numpy==1.25.1
opencv-python==4.8.0.76
pillow==9.5.0
requests==2.25.1
streamlit==1.25.0
pyyaml